Implement cycle decomposition livelock algorithm
Instead of treating all the threads as a single interacting unit, this
algorithm instead leverages the observation that most commonly occurring
livelock scenarios decompose into multiple independent cycles.

Thus, by observing each independent cycle in isolation, the algorithm
aims to make livelock witnesses short and easy to understand while also
increasing the efficiency of livelock detection.

+++ src/mcmini_private.cpp
@@ -660,7 +660,7 @@ mc_search_dpor_branch_with_thread(const tid_t backtrackThread)
         nextTransition = nullptr;
       }
       if (getenv(ENV_CHECK_FOR_LIVELOCK) && nextTransition == nullptr) {
-        uint64_t increasedDepth = LLOCK_INCREASED_MAX_TRANSITIONS_DEPTH;
+        int increasedDepth = LLOCK_INCREASED_MAX_TRANSITIONS_DEPTH;
         if (getenv(ENV_MAX_LIVELOCK_CYCLE_LIMIT)) {
           increasedDepth =
             strtoul(getenv(ENV_MAX_LIVELOCK_CYCLE_LIMIT), nullptr, 10);
